hotfix: ffmpeg encode timeout safety. read notes
After doing a decent amount of debugging. It seems there is a very rare case within the ffmpeg encoding process where it completely freezes without failing or finishing.
Despite hooks to catch errors and escalate them upwards it seems to be a definite issue.
So, for now I am introducing an a max cap ffmpeg can stale without making progress.